keep the input data type

+++ clawdia/lib.py
@@ -182,7 +182,7 @@ def extract_patches(signals, *, patch_size, n_patches=None, random_state=None,
             f" the maximum number of patches that can be extracted ({max_patches})."
         )
     
-    patches = np.empty((patch_size, n_patches), order='F')
+    patches = np.empty((patch_size, n_patches), dtype=signals.dtype, order='F')
 
     # Extract all possible patches.
     if n_patches == max_patches:
@@ -219,7 +219,7 @@ def reconstruct_from_patches_1d(patches, step):
     l_patches, n_patches = patches.shape
     total_len = (n_patches - 1) * step + l_patches
     
-    reconstructed = np.zeros(total_len)
+    reconstructed = np.zeros(total_len, dtype=patches.dtype)
     normalizer = np.zeros_like(reconstructed)
     for i in range(n_patches):
         reconstructed[i*step:i*step+l_patches] += patches[:,i]
